Create LeastConnection(naive).cpp
Least Connection dynamic load balancing algorithm is a technique where the next task is assigned to the processor with the least number of active tasks or connections at that moment. 

for this naive approach i.e. without using heap/priority queue it is taking O(P) for assigning task and O(T) for overall task which make its time complexity to be O(T * P)